requirements.txt
appdirs==1.4.4
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
dataclasses==0.8
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
nose==1.3.7
protobuf==3.19.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
webdrivermanager==0.10.0
Werkzeug==2.0.2
zipp==3.6.0
